requirements.txt
transformers==4.40.2
tokenizers==0.19.1
torch==2.2.2
scikit-learn==1.4.2
pyyaml==6.0.1
wandb==0.17.0
dvc[s3]==3.51.0
awscli==1.32.98
docker==7.0.0
gradio==4.28.3
boto3==1.34.98
datasets==2.19.1
evaluate==0.4.1
accelerate==0.29.2
sentencepiece==0.1.99